+++ integrationtests/webtransport_test.go
@@ -1,4 +1,4 @@
-package webtransport_test
+package integrationtests
 
 import (
 	"context"
@@ -10,6 +10,7 @@ import (
 	"io"
 	"net"
 	"net/http"
+	"os"
 	"strconv"
 	"strings"
 	"sync"
@@ -17,6 +18,7 @@ import (
 	"time"
 
 	"github.com/quic-go/webtransport-go"
+	"github.com/quic-go/webtransport-go/internal/testdata"
 
 	"github.com/quic-go/quic-go"
 	"github.com/quic-go/quic-go/http3"
@@ -25,6 +27,13 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+func scaleDuration(d time.Duration) time.Duration {
+	if os.Getenv("CI") != "" {
+		return 5 * d
+	}
+	return d
+}
 
 func runServer(t *testing.T, s *webtransport.Server) (addr *net.UDPAddr, close func()) {
 	laddr, err := net.ResolveUDPAddr("udp", "localhost:0")
@@ -48,7 +57,7 @@ func runServer(t *testing.T, s *webtransport.Server) (addr *net.UDPAddr, close f
 func establishSession(t *testing.T, handler func(*webtransport.Session)) (sess *webtransport.Session, close func()) {
 	s := &webtransport.Server{
 		H3: &http3.Server{
-			TLSConfig: webtransport.TLSConf,
+			TLSConfig: testdata.TLSConf,
 			QUICConfig: &quic.Config{
 				EnableDatagrams:                  true,
 				EnableStreamResetPartialDelivery: true,
@@ -60,7 +69,7 @@ func establishSession(t *testing.T, handler func(*webtransport.Session)) (sess *
 
 	addr, closeServer := runServer(t, s)
 	d := webtransport.Dialer{
-		TLSClientConfig: &tls.Config{RootCAs: webtransport.CertPool},
+		TLSClientConfig: &tls.Config{RootCAs: testdata.CertPool},
 		QUICConfig: &quic.Config{
 			EnableDatagrams:                  true,
 			EnableStreamResetPartialDelivery: true,
@@ -169,7 +178,7 @@ func TestApplicationProtocolNegotiationErrors(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			s := &webtransport.Server{
 				H3: &http3.Server{
-					TLSConfig: webtransport.TLSConf,
+					TLSConfig: testdata.TLSConf,
 					QUICConfig: &quic.Config{
 						EnableDatagrams:                  true,
 						EnableStreamResetPartialDelivery: true,
@@ -189,7 +198,7 @@ func TestApplicationProtocolNegotiationErrors(t *testing.T) {
 			defer closeServer()
 
 			d := webtransport.Dialer{
-				TLSClientConfig: &tls.Config{RootCAs: webtransport.CertPool},
+				TLSClientConfig: &tls.Config{RootCAs: testdata.CertPool},
 				QUICConfig: &quic.Config{
 					EnableDatagrams:                  true,
 					EnableStreamResetPartialDelivery: true,
@@ -217,7 +226,7 @@ func testApplicationProtocolNegotiation(t *testing.T, clientProtocols, serverPro
 	s := &webtransport.Server{
 		ApplicationProtocols: serverProtocols,
 		H3: &http3.Server{
-			TLSConfig: webtransport.TLSConf,
+			TLSConfig: testdata.TLSConf,
 			QUICConfig: &quic.Config{
 				EnableDatagrams:                  true,
 				EnableStreamResetPartialDelivery: true,
@@ -234,7 +243,7 @@ func testApplicationProtocolNegotiation(t *testing.T, clientProtocols, serverPro
 	addr, closeServer := runServer(t, s)
 	defer closeServer()
 	d := webtransport.Dialer{
-		TLSClientConfig: &tls.Config{RootCAs: webtransport.CertPool},
+		TLSClientConfig: &tls.Config{RootCAs: testdata.CertPool},
 		QUICConfig: &quic.Config{
 			EnableDatagrams:                  true,
 			EnableStreamResetPartialDelivery: true,
@@ -395,7 +404,7 @@ func TestUnidirectionalStreams(t *testing.T) {
 func TestMultipleClients(t *testing.T) {
 	const numClients = 5
 	s := &webtransport.Server{
-		H3: &http3.Server{TLSConfig: webtransport.TLSConf},
+		H3: &http3.Server{TLSConfig: testdata.TLSConf},
 	}
 	defer s.Close()
 	addHandler(t, s, newEchoHandler(t))
@@ -409,7 +418,7 @@ func TestMultipleClients(t *testing.T) {
 		go func() {
 			defer wg.Done()
 			d := webtransport.Dialer{
-				TLSClientConfig: &tls.Config{RootCAs: webtransport.CertPool},
+				TLSClientConfig: &tls.Config{RootCAs: testdata.CertPool},
 				QUICConfig: &quic.Config{
 					EnableDatagrams:                  true,
 					EnableStreamResetPartialDelivery: true,
@@ -592,7 +601,7 @@ func TestCheckOrigin(t *testing.T) {
 	for _, tc := range tcs {
 		t.Run(tc.Name, func(t *testing.T) {
 			s := &webtransport.Server{
-				H3:          &http3.Server{TLSConfig: webtransport.TLSConf},
+				H3:          &http3.Server{TLSConfig: testdata.TLSConf},
 				CheckOrigin: tc.CheckOrigin,
 			}
 			defer s.Close()
@@ -602,7 +611,7 @@ func TestCheckOrigin(t *testing.T) {
 			defer closeServer()
 
 			d := webtransport.Dialer{
-				TLSClientConfig: &tls.Config{RootCAs: webtransport.CertPool},
+				TLSClientConfig: &tls.Config{RootCAs: testdata.CertPool},
 				QUICConfig:      &quic.Config{Tracer: qlog.DefaultConnectionTracer, EnableDatagrams: true, EnableStreamResetPartialDelivery: true},
 			}
 			defer d.Close()
@@ -797,7 +806,7 @@ func TestSessionContextValues(t *testing.T) {
 
 	s := &webtransport.Server{
 		H3: &http3.Server{
-			TLSConfig:  webtransport.TLSConf,
+			TLSConfig:  testdata.TLSConf,
 			QUICConfig: &quic.Config{Tracer: qlog.DefaultConnectionTracer, EnableDatagrams: true, EnableStreamResetPartialDelivery: true},
 		},
 	}
@@ -821,7 +830,7 @@ func TestSessionContextValues(t *testing.T) {
 	defer closeServer()
 
 	d := webtransport.Dialer{
-		TLSClientConfig: &tls.Config{RootCAs: webtransport.CertPool},
+		TLSClientConfig: &tls.Config{RootCAs: testdata.CertPool},
 		QUICConfig:      &quic.Config{Tracer: qlog.DefaultConnectionTracer, EnableDatagrams: true, EnableStreamResetPartialDelivery: true},
 	}
 	defer d.Close()
