+++ test_helpers_test.go
@@ -1,102 +1,18 @@
 package webtransport
 
 import (
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/tls"
-	"crypto/x509"
-	"crypto/x509/pkix"
-	"log"
-	"math/big"
 	"net/http"
 	"net/url"
 	"testing"
-	"time"
 
-	"github.com/quic-go/quic-go/http3"
+	"github.com/quic-go/webtransport-go/internal/testdata"
 	"github.com/stretchr/testify/require"
 )
 
 var (
-	TLSConf  *tls.Config
-	CertPool *x509.CertPool
+	TLSConf  = testdata.TLSConf
+	CertPool = testdata.CertPool
 )
-
-// use a very long validity period to cover the synthetic clock used in synctest
-var (
-	notBefore = time.Date(1990, 1, 1, 0, 0, 0, 0, time.UTC)
-	notAfter  = time.Date(2100, 1, 1, 0, 0, 0, 0, time.UTC)
-)
-
-func init() {
-	ca, caPrivateKey, err := generateCA()
-	if err != nil {
-		log.Fatal("failed to generate CA certificate:", err)
-	}
-	leafCert, leafPrivateKey, err := generateLeafCert(ca, caPrivateKey)
-	if err != nil {
-		log.Fatal("failed to generate leaf certificate:", err)
-	}
-	CertPool = x509.NewCertPool()
-	CertPool.AddCert(ca)
-	TLSConf = &tls.Config{
-		Certificates: []tls.Certificate{{
-			Certificate: [][]byte{leafCert.Raw},
-			PrivateKey:  leafPrivateKey,
-		}},
-		NextProtos: []string{http3.NextProtoH3},
-	}
-}
-
-func generateCA() (*x509.Certificate, *rsa.PrivateKey, error) {
-	certTempl := &x509.Certificate{
-		SerialNumber:          big.NewInt(2019),
-		Subject:               pkix.Name{},
-		NotBefore:             notBefore,
-		NotAfter:              notAfter,
-		IsCA:                  true,
-		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth, x509.ExtKeyUsageServerAuth},
-		KeyUsage:              x509.KeyUsageDigitalSignature | x509.KeyUsageCertSign,
-		BasicConstraintsValid: true,
-	}
-	caPrivateKey, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		return nil, nil, err
-	}
-	caBytes, err := x509.CreateCertificate(rand.Reader, certTempl, certTempl, &caPrivateKey.PublicKey, caPrivateKey)
-	if err != nil {
-		return nil, nil, err
-	}
-	ca, err := x509.ParseCertificate(caBytes)
-	if err != nil {
-		return nil, nil, err
-	}
-	return ca, caPrivateKey, nil
-}
-
-func generateLeafCert(ca *x509.Certificate, caPrivateKey *rsa.PrivateKey) (*x509.Certificate, *rsa.PrivateKey, error) {
-	certTempl := &x509.Certificate{
-		SerialNumber: big.NewInt(1),
-		DNSNames:     []string{"localhost"},
-		NotBefore:    notBefore,
-		NotAfter:     notAfter,
-		ExtKeyUsage:  []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth, x509.ExtKeyUsageServerAuth},
-		KeyUsage:     x509.KeyUsageDigitalSignature,
-	}
-	privKey, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		return nil, nil, err
-	}
-	certBytes, err := x509.CreateCertificate(rand.Reader, certTempl, ca, &privKey.PublicKey, caPrivateKey)
-	if err != nil {
-		return nil, nil, err
-	}
-	cert, err := x509.ParseCertificate(certBytes)
-	if err != nil {
-		return nil, nil, err
-	}
-	return cert, privKey, nil
-}
 
 func NewWebTransportRequest(t *testing.T, addr string) *http.Request {
 	t.Helper()
